+++ src/LLP/types/tritium.h
@@ -177,7 +177,7 @@ namespace LLP
                     if(Read(DATA, DATA.size()) == DATA.size())
                     {
                         INCOMING.DATA.insert(INCOMING.DATA.end(), DATA.begin(), DATA.end());
-                        Event(EVENT_PACKET, DATA.size());
+                        Event(EVENT_PACKET, static_cast<uint32_t>(DATA.size()));
                     }
                 }
             }
